requirements.txt
django
djangorestframework
python-dotenv
environs
django_geopostcodes 
icalendar
pytz
requests